requirements.txt
praw==7.2.0
python-decouple==3.4
requests==2.25.1